+++ scripts/codex.js
@@ -1,0 +1,30 @@
+const codex = [
+  {
+    name: "Materials & Crafting",
+    text: "This part touches upon all the materials and crafting recipes found in the smithy.",
+    content: [
+      {
+        name: "Monster Core",
+        text: "Monster cores are materials used in crafting that drop from powerful monsters, such as the §/$Y/Orc Berserker§.",
+        image: "monster_core.png"
+      },
+      {
+        name: "Iron Ingot",
+        text: "Iron Ingots are the most basic crafting material, easily bought from the store.",
+        image: "iron_ingot.png"
+      },
+      {
+        name: "Steel Ingot",
+        text: "Steel Ingots are more advanced crafting material than iron ingots, but can still be bought from the store.",
+        image: "steel_ingot.png"
+      }
+    ]
+  },
+  {
+    name: "Enemies",
+    text: "You can view the statistics and trivia of all the enemies you have defeated here.",
+    content: {
+      
+    }
+  }
+]